fix: trim GitHub extension test-only hooks
Remove an unused watched-repo change callback and keep the static config test helper out of the production GitHub extension module.
